% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{cam_move}
\alias{cam_move}
\title{Move files from the camera folders to the location folders and rename images
automatically}
\usage{
cam_move(project = NULL, overwrite = F, date_format = NULL, tz = "UTC")
}
\arguments{
\item{project}{character. Directory to the project folder.}

\item{overwrite}{logical. In case the renamed file already exists in the
locations folders. Should it be overwritten?}

\item{date_format}{character. Format in which the dates in the
'Deployments.csv' file are entered. If not provided, it will be guessed.}

\item{tz}{character. Time zone of the dates. Set to UTC by default.}
}
\description{
This function takes the images assigned to each camera and copies them to the
folder of the location at which the camera was setup. For this, the
'Deployments.csv' file has to be duly filled out.
}
\examples{
Coming
}
