% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{cam_project}
\alias{cam_project}
\title{Prepare folder Structure}
\usage{
cam_project(name = NULL, dir = NULL, cameras = NULL, locations = NULL)
}
\arguments{
\item{name}{character. Name of the project. This will become the name of the
main folder}

\item{dir}{character. directory to the folder where the project should be
located}

\item{cameras}{character vector. CameraIDs that will be used for the study.}

\item{locations}{(OPTIONAL) character vector. LocationIDs at which the
cameras will be placed.}
}
\description{
Function to create all necessary folders to work with the cameratrapr package
}
\examples{
Coming
}
