% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{chainLengths}
\alias{chainLengths}
\title{DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD}
\usage{
chainLengths(chain)
}
\value{

}
\description{
Calculate footprint chain length and number of steps
}
\examples{
attach(chainDB)
footprintlength = chainLength(chainDB); footprintlength

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
