% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{dailyRhythm}
\alias{dailyRhythm}
\title{plot the daily rhythm of species' activity}
\usage{
dailyRhythm(x)
}
\arguments{
\item{x}{A data.frame with column names "Lon", "Lat", "Group_size", "Date", "Time"}
}
\value{

}
\description{
This function plots the curves of probability density of animal activity (density of picture time) in 24 hours
}
\examples{

 attach(trapresult)
 dailyRhythm(trapresult)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
