% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package.r
\name{geo2proj}
\alias{geo2proj}
\title{Convertr longitude & latitude from geographic to projected coordinate system}
\usage{
geo2proj(data, geo = NULL, proj = NULL)
}
\arguments{
\item{data}{data.frame with column names "Lon" & "Lat"}

\item{geo}{geographic coordinate system, e.g. "+proj=longlat +datum=WGS84"}

\item{proj}{projection coordinate system, e.g. "+proj=utm +zone=30 ellps=WGS84"}
}
\value{
Return a data.frame with column names "Lon" & "Lat". The unit of "Lon" & "Lat" is meter now.
}
\description{
This function converts longitude & latitude from geographic to projected coordinate system.
 Users can specify their coordinate reference systems, or leave them as default. The default setting will
 using "WGS83" as geographic coordinates system and "UTM" as the projection coordinates system. The "zone"
 for UTM projection system will be determined using the centriod of points (longitude & latitude).
}
\examples{

 camera.geo <- data.frame(Lon = rnorm(10, 120, 10), Lat = rnorm(10, 30, 5))
 par(mfrow = c(1, 2))
 with(camera.geo, plot(Lon, Lat, xlab = "Longitude", ylab = "Latitude", main = "Geographic coordinate system"))

 camera.proj <- geo2proj(camera.geo)
 with(camera.proj, plot(Lon, Lat, xlab = "Longitude", ylab = "Latitude", main = "Geographic coordinate system"))
}
\author{
Huidong Tian (tienhuitung@gmail.com)
}
