% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{homeRange}
\alias{homeRange}
\title{HHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHH}
\usage{
homeRange(chain)
}
\value{

}
\description{
Calculate home ranges based on footprint chains
}
\examples{
attach(footprintchain)
HOME = homeRange(footprintchain); HOME

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
