% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{homeRanges}
\alias{homeRanges}
\title{HHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHH}
\usage{
homeRanges(chain)
}
\value{

}
\description{
Calculate home ranges based on footprint chains
}
\examples{

attach(chainDB)
HOME = homeRange(chainDB);HOME

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
