% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package.r
\name{moveBias}
\alias{moveBias}
\title{Show how likely the animal would turn during moving}
\usage{
moveBias(chain, scale)
}
\arguments{
\item{chain}{A data.frame with column names "Lon", "Lat", and "Date", in the order of recording time
at the interval of one second}

\item{scale}{A value defining the step length}
}
\value{

}
\description{
Plot the probability density of step length, and probability density of turning angle between two steps.
 The step length can be defined by "scale" as the distance between every one second, two seconds, etc.
}
\examples{

 attach(footprintchain)
 moveBias(footprintchain, scale=2)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
