% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package.r
\name{plotCamtrap}
\alias{plotCamtrap}
\title{plot camera trapping results, showing the locations of the cameras and the number of pictures}
\usage{
plotCamtrap(x, circle.size = 0.2, point.scatter = 5)
}
\arguments{
\item{x}{A data.frame with column names "Lon", "Lat", "Group_size", "Date", "Time"}

\item{circle.size}{A value controls the size the circles. The defualt value is 0.2.}

\item{point.scatter}{A value controls the distance between every point (representing a picture)
at the same camera. The defualt value is 5.}
}
\value{

}
\description{
This function shows the locations of each camera in a camera trap grid, the number of pictures
 taken by each camera for one species.
}
\examples{

 attach(trapresult)
 plotCamtrap(trapresult)
 plotCamtrap(trapresult, circle.size = .5, point.scatter = 10)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
