% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{plotFootprint}
\alias{plotFootprint}
\title{Plot the trajectory of footprint chain}
\usage{
plotFootprint(chain)
}
\arguments{
\item{chain}{A data.frame with column names "Lon", "Lat", and "Date", in the order of recording time
at the interval of one second}
}
\value{

}
\description{
Plot the trajectory of footprint chain, highlighting the starting and ending points.
}
\examples{

 attach(footprintchain)
 plotFootprint(footprintchain)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
