% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{predictCamtrap}
\alias{predictCamtrap}
\title{Predict number of individuals within the range of camera trap grid}
\usage{
predictCamtrap(simu, x, plot = F)
}
\arguments{
\item{simu}{The result of function simuCamtrap().}

\item{x}{A data.frame with column names "Lon", "Lat", "Group_size", "Date", "Time"}

\item{plot}{a boolean variable, if TRUE, plot the probability density of estimated animal abundance}
}
\value{
The mean value, 95\% confidence intervals of the predicted animal abundance, based on
a vector of the predicted animal abundance from 1000 random forest trees.
}
\description{
This function matches the real camera trap result with pseudo camera trap result simulated for a series number of individuals,
to find the best fit of animal abundance with real camera trap result.
}
\examples{

 attach(trapresult)
 # sim.out = simuCamtrap(trapresult, ind = 10, iteration = 2) # need a few minutes
 predictCamtrap(sim.out, trapresult, plot=T)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
