% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{Rowcliffe}
\alias{Rowcliffe}
\title{Estimate number of individuals using Rowcliffe's equation (Rowcliffe 2008)}
\usage{
Rowcliffe(x, r, theta, v, duration)
}
\arguments{
\item{x}{A data.frame with column names "Lon", "Lat", "Group_size", "Date", "Time".}

\item{r}{Detection range (km) of the camera.}

\item{theta}{Detection angle of the camera}

\item{v}{Mean velocity (km/h) of the animal.}

\item{duration}{Days for camera trapping.}
}
\value{

}
\description{
Calculate animal density using Rowcliffe's random encounter models.
}
\examples{

attach(trapresult)
Rowcliffe(trapresult, r=0.02, theta = 40, v = 2, duration = 40) # unit: km

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
