% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{video2pic}
\alias{video2pic}
\title{Convert video to jpeg}
\usage{
video2pic(file = NULL, outdir = NULL, fps = NULL, overwrite = F)
}
\arguments{
\item{outdir}{directory to the folder where the converted file(s) should be
stored. By default, this is the working directory.}

\item{fps}{numeric frames per second that should be stored}

\item{filepath}{filepath(s) to the video file(s) that need to be converted}

\item{logical}{overwrite should frames that already exist in the output directory be overwritten?}
}
\description{
Function to convert a video into frames
}
\examples{
Coming
}
