% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{chainLength}
\alias{chainLength}
\title{DDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDDD}
\usage{
chainLength(chain)
}
\value{

}
\description{
Calculate footprint chain length and number of steps
}
\examples{
attach(footprintchain)
footprintlength = chainLength(footprintchain); footprintlength

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
