% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{homeRange}
\alias{homeRange}
\title{HHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHHH}
\usage{
homeRange(chain)
}
\value{

}
\description{
Calculate home ranges based on footprint chains
}
\examples{

 chain = read.csv("D:/database/cameratrap/footprintchain.csv", header=T);head(chain)
 chain = chain[1:3000,]
 chain = droplevels(chain)

 attach(chain)
 HOME = homeRange(chain);HOME

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
