% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera trap package E.r
\name{simuCamtrap}
\alias{simuCamtrap}
\title{Simulate animal movement within the range of the camera trap grid and obtain the pseudo camera trap result}
\usage{
simuCamtrap(
  x,
  detect = 50,
  bearing = runif(camera.N, 0, 2 * pi),
  step.N = 5000,
  step.V = 2,
  step.L = 10,
  bias = 30/360 * 2 * pi,
  range = 4000,
  ind = 10,
  iteration = 3
)
}
\arguments{
\item{x}{A data.frame with column names "Lon", "Lat", "Group_size", "Date", "Time"}

\item{detect}{The detection radias (m) of a camera.}

\item{bearing}{The bearing direction of a camera.}

\item{step.N}{The number of steps the animal walks during the camera trapping.}

\item{step.V}{The standard diviation of the step length}

\item{step.L}{The mean step length (m) of the animal.}

\item{bias}{The standard diviation of the changing angle (degree) between two steps}

\item{range}{Maximum distance (m) the animal moves from the original site.}

\item{ind}{The number of individuals that are simulated.}

\item{iteration}{The number of simulations.}
}
\value{
A dataframe with the first column to be the number of individuals, and the rest columns
are number of pictures (simulated) for each camera
}
\description{
Correlated random walk of the target animal is simulated within the range of camera trap grid,
using the distributions of step length, turning angles, and size of home range from footprint chain data.
The simulated movement of the default 1-10 individuals generate pseudo camera trap data, which are matched
with the real data using the random forest algorithm, in order to find the best fit of animal abundance
among the abundance from 1 to 10 taken by each camera for one species. Such simulation can be repeated for
several times defined by number of iteration.
}
\examples{

par(mfrow = c(1, 2))
# maximum number of individuals in the camera grid is 10 (ind=10)
sim.out = simuCamtrap(trapresult, ind = 10, iteration = 2) # more iterations are expected for higher prediction accuracy

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
